mvi-kotlin/app/src/main/kotlin/io/github/ruieduardosoares/android/mvi/kotlin/presenter/StateStreamToStateRelayBinder.kt
/**
 * MIT License
 *
 * Copyright (c) [2021] [Rui Eduardo Soares]
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package io.github.ruieduardosoares.android.mvi.kotlin.presenter

import io.reactivex.rxjava3.core.Observable
import io.reactivex.rxjava3.disposables.Disposable
import io.reactivex.rxjava3.subjects.BehaviorSubject

internal class StateStreamToStateRelayBinder<S : Any>(
    private val mViewRenderRelaySubject: BehaviorSubject<S>
) {

    private var mDisposable: Disposable? = null

    fun bind(stateStream: Observable<S>) {
        mDisposable?.let { disposable -> if (disposable.isDisposed.not()) return }
        mDisposable =
            stateStream.subscribe(
                mViewRenderRelaySubject::onNext,
                mViewRenderRelaySubject::onError,
                mViewRenderRelaySubject::onComplete
            )
    }

    fun dispose() {
        mDisposable = mDisposable?.let {
            it.dispose()
            null
        }
    }
}
